restructuring, comments from review
