+++ pytorch_to_returnn/import_wrapper/torch_wrappers/__init__.py
@@ -2,3 +2,4 @@
 from .tensor import WrappedTorchTensor
 from .parameter import WrappedTorchParameter
 from .module import WrappedModuleBase
+from .function import Functional
